[Refactor] move more to base
