+++ be/src/exec/pipeline/exchange/mem_limited_chunk_queue.h
@@ -18,10 +18,10 @@
 #include <queue>
 
 #include "base/bit/bit_mask.h"
+#include "base/concurrency/spinlock.h"
 #include "exec/pipeline/exchange/multi_cast_local_exchange.h"
 #include "serde/encode_context.h"
 #include "util/phmap/phmap.h"
-#include "util/spinlock.h"
 
 namespace starrocks::pipeline {
