+++ be/src/exec/pipeline/fetch_processor.h
@@ -18,6 +18,7 @@
 #include <queue>
 #include <vector>
 
+#include "base/concurrency/bthread_shared_mutex.h"
 #include "bthread/mutex.h"
 #include "column/column.h"
 #include "column/vectorized_fwd.h"
@@ -27,7 +28,6 @@
 #include "exec/pipeline/lookup_request.h"
 #include "exec/sorting/sort_permute.h"
 #include "runtime/descriptors.h"
-#include "util/bthreads/bthread_shared_mutex.h"
 #include "util/phmap/phmap.h"
 #include "util/raw_container.h"
