+++ be/src/exec/pipeline/pipeline_driver_queue.h
@@ -16,10 +16,10 @@
 
 #include <queue>
 
+#include "base/concurrency/moodycamel/concurrentqueue.h"
 #include "base/utility/factory_method.h"
 #include "exec/pipeline/pipeline_driver.h"
 #include "exec/workgroup/work_group_fwd.h"
-#include "util/moodycamel/concurrentqueue.h"
 
 namespace starrocks::pipeline {
