+++ be/src/exec/pipeline/query_context.h
@@ -20,6 +20,7 @@
 #include <optional>
 #include <unordered_map>
 
+#include "base/concurrency/spinlock.h"
 #include "base/hash/hash.h"
 #include "exec/pipeline/fragment_context.h"
 #include "exec/pipeline/pipeline_fwd.h"
@@ -33,7 +34,6 @@
 #include "runtime/runtime_state.h"
 #include "util/debug/query_trace.h"
 #include "util/hash_util.hpp"
-#include "util/spinlock.h"
 #include "util/time.h"
 
 namespace starrocks {
