+++ be/src/exec/pipeline/scan/scan_operator.h
@@ -15,6 +15,7 @@
 #pragma once
 
 #include "base/concurrency/race_detect.h"
+#include "base/concurrency/spinlock.h"
 #include "exec/exec_node.h"
 #include "exec/pipeline/pipeline_fwd.h"
 #include "exec/pipeline/scan/balanced_chunk_buffer.h"
@@ -23,7 +24,6 @@
 #include "exec/query_cache/cache_operator.h"
 #include "exec/query_cache/lane_arbiter.h"
 #include "exec/workgroup/work_group_fwd.h"
-#include "util/spinlock.h"
 
 namespace starrocks {
