+++ be/src/exec/pipeline/sink/mysql_table_sink_operator.cpp
@@ -22,10 +22,10 @@
 #ifndef __APPLE__
 #include "runtime/mysql_table_writer.h"
 #endif
+#include "base/concurrency/spinlock.h"
 #include "base/utility/defer_op.h"
 #include "runtime/runtime_state.h"
 #include "udf/java/utils.h"
-#include "util/spinlock.h"
 
 namespace starrocks::pipeline {
