+++ be/src/runtime/batch_write/batch_write_mgr.h
@@ -16,11 +16,11 @@
 
 #include <unordered_map>
 
+#include "base/concurrency/bthread_shared_mutex.h"
 #include "common/statusor.h"
 #include "runtime/batch_write/isomorphic_batch_write.h"
 #include "runtime/batch_write/txn_state_cache.h"
 #include "runtime/stream_load/stream_load_context.h"
-#include "util/bthreads/bthread_shared_mutex.h"
 #include "util/bthreads/executor.h"
 
 namespace brpc {
