+++ be/src/runtime/lake_tablets_channel.cpp
@@ -20,6 +20,7 @@
 #include <unordered_map>
 #include <vector>
 
+#include "base/concurrency/bthread_shared_mutex.h"
 #include "column/chunk.h"
 #include "common/closure_guard.h"
 #include "common/compiler_util.h"
@@ -41,7 +42,6 @@
 #include "storage/memtable.h"
 #include "storage/memtable_flush_executor.h"
 #include "storage/storage_engine.h"
-#include "util/bthreads/bthread_shared_mutex.h"
 #include "util/compression/block_compression.h"
 #include "util/countdown_latch.h"
 #include "util/runtime_profile.h"
