+++ be/src/runtime/lookup_stream_mgr.h
@@ -19,6 +19,7 @@
 #include <mutex>
 #include <shared_mutex>
 
+#include "base/concurrency/moodycamel/concurrentqueue.h"
 #include "bthread/mutex.h"
 #include "column/vectorized_fwd.h"
 #include "common/global_types.h"
@@ -29,7 +30,6 @@
 #include "runtime/descriptors.h"
 #include "runtime/result_queue_mgr.h"
 #include "runtime/runtime_state.h"
-#include "util/moodycamel/concurrentqueue.h"
 #include "util/phmap/phmap.h"
 #include "util/phmap/phmap_base.h"
 #include "util/phmap/phmap_utils.h"
