+++ be/src/runtime/record_batch_queue.h
@@ -17,9 +17,8 @@
 
 #pragma once
 
-#include <util/spinlock.h>
-
 #include "base/concurrency/blocking_queue.hpp"
+#include "base/concurrency/spinlock.h"
 #include "common/status.h"
 
 namespace arrow {
