+++ be/src/runtime/user_function_cache.cpp
@@ -40,6 +40,7 @@
 #include <utility>
 #include <vector>
 
+#include "base/concurrency/spinlock.h"
 #include "common/config.h"
 #include "common/status.h"
 #include "fmt/compile.h"
@@ -48,7 +49,6 @@
 #include "gutil/strings/split.h"
 #include "util/download_util.h"
 #include "util/dynamic_util.h"
-#include "util/spinlock.h"
 
 namespace starrocks {
