+++ be/src/storage/memtable_flush_executor.h
@@ -38,10 +38,10 @@
 #include <memory>
 #include <vector>
 
+#include "base/concurrency/spinlock.h"
 #include "common/status.h"
 #include "storage/memtable.h"
 #include "storage/olap_define.h"
-#include "util/spinlock.h"
 #include "util/threadpool.h"
 
 namespace starrocks {
