+++ be/src/storage/rowset/unique_rowset_id_generator.cpp
@@ -34,7 +34,7 @@
 
 #include "storage/rowset/unique_rowset_id_generator.h"
 
-#include "util/spinlock.h"
+#include "base/concurrency/spinlock.h"
 #include "util/starrocks_metrics.h"
 #include "util/uid_util.h"
