+++ be/src/storage/rowset/unique_rowset_id_generator.h
@@ -36,8 +36,8 @@
 
 #include <atomic>
 
+#include "base/concurrency/spinlock.h"
 #include "storage/rowset/rowset_id_generator.h"
-#include "util/spinlock.h"
 #include "util/uid_util.h"
 
 namespace starrocks {
