+++ be/src/storage/segment_replicate_executor.h
@@ -18,12 +18,12 @@
 #include <memory>
 #include <vector>
 
+#include "base/concurrency/spinlock.h"
 #include "common/status.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "storage/olap_define.h"
 #include "util/internal_service_recoverable_stub.h"
 #include "util/reusable_closure.h"
-#include "util/spinlock.h"
 #include "util/threadpool.h"
 
 namespace starrocks {
