+++ be/src/util/trace.h
@@ -24,13 +24,13 @@
 #include <utility>
 #include <vector>
 
+#include "base/concurrency/spinlock.h"
 #include "gutil/macros.h"
 #include "gutil/ref_counted.h"
 #include "gutil/strings/stringpiece.h"
 #include "gutil/strings/substitute.h"
 #include "gutil/threading/thread_collision_warner.h"
 #include "gutil/walltime.h"
-#include "util/spinlock.h"
 #include "util/trace_metrics.h"
 
 namespace starrocks {
